requirements.txt
asgiref==3.8.1
attrs==24.2.0
backports.zoneinfo==0.2.1
coverage==7.6.1
Django==4.2.16
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.2
importlib-resources==6.4.5
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.2
pkgutil-resolve-name==1.3.10
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
sqlparse==0.5.1
tomli==2.1.0
typing-extensions==4.12.2
uritemplate==4.1.1
zipp==3.20.2
